+++ src/bin/tests/test_multicast_receiver.cpp
@@ -3,7 +3,7 @@
 #include <boost/asio.hpp>
 #include "boost/bind.hpp"
 
-#include "travesim_adapters/multicast_receiver.hh"
+#include "travesim_adapters/udp/multicast_receiver.hh"
 
 #define BUFFER_SIZE 1024U
